Move error page to /pages directory
